refactor(Infra): GCP cache usage
The GCP component is about to be refactored to allow the usage of APCU for cache storage.
By default, the configuration is updated and in place, the Extension should be updated.
The Writer and Repository are updated to allow the usage of both Redis & APCU connectors.

This refactoring is related to https://github.com/GuikProd/MarketReminder/issues/14
This refactoring is related to https://github.com/GuikProd/MarketReminder/issues/15
This refactoring is related to https://github.com/GuikProd/MarketReminder/issues/16